Make side-effect.js smaller
